@@ -1,4 +1,4 @@
-<?php
+<?php declare(strict_types=1);
 /**
  * @author Igor Timoshenkov [it@campoint.net]
  * @started: 03.09.2018 9:21
@@ -14,35 +14,12 @@ use Psr\Http\Message\ResponseInterface;
  */
 class MessageSentReport implements \JsonSerializable
 {
-    /**
-     * @var boolean
-     */
-    protected $success;
-
-    /**
-     * @var RequestInterface
-     */
-    protected $request;
-
-    /**
-     * @var ResponseInterface | null
-     */
-    protected $response;
-
-    /**
-     * @var string
-     */
-    protected $reason;
-
-    /**
-     * @param string $reason
-     */
-    public function __construct(RequestInterface $request, ?ResponseInterface $response = null, bool $success = true, $reason = 'OK')
-    {
-        $this->request  = $request;
-        $this->response = $response;
-        $this->success  = $success;
-        $this->reason   = $reason;
+    public function __construct(
+        protected RequestInterface $request,
+        protected ?ResponseInterface $response = null,
+        protected bool $success = true,
+        protected string $reason = 'OK'
+    ) {
     }
 
     public function isSuccess(): bool
@@ -110,11 +87,7 @@ class MessageSentReport implements \JsonSerializable
 
     public function getResponseContent(): ?string
     {
-        if (!$this->response) {
-            return null;
-        }
-
-        return $this->response->getBody()->getContents();
+        return $this->response?->getBody()->getContents();
     }
 
     public function jsonSerialize(): array
