@@ -3,7 +3,7 @@
         'name' => '__root__',
         'pretty_version' => 'dev-master',
         'version' => 'dev-master',
-        'reference' => '6f6461b4a24316d70856362b6136992572e05b32',
+        'reference' => 'dafe4e4d98ef3d45aa1c838703676812016db738',
         'type' => 'project',
         'install_path' => __DIR__ . '/../',
         'aliases' => array(),
@@ -13,7 +13,7 @@
         '__root__' => array(
             'pretty_version' => 'dev-master',
             'version' => 'dev-master',
-            'reference' => '6f6461b4a24316d70856362b6136992572e05b32',
+            'reference' => 'dafe4e4d98ef3d45aa1c838703676812016db738',
             'type' => 'project',
             'install_path' => __DIR__ . '/../',
             'aliases' => array(),
@@ -128,9 +128,9 @@
             'dev_requirement' => false,
         ),
         'minishlink/web-push' => array(
-            'pretty_version' => 'v8.0.0',
-            'version' => '8.0.0.0',
-            'reference' => 'ec034f1e287cd1e74235e349bd017d71a61e9d8d',
+            'pretty_version' => 'v9.0.0-rc2',
+            'version' => '9.0.0.0-RC2',
+            'reference' => '9d36211c435baecded11d7a227f5caa098f52f80',
             'type' => 'library',
             'install_path' => __DIR__ . '/../minishlink/web-push',
             'aliases' => array(),
